pr comments and restructure files to make propertyline distinction more clear
