+++ packages/dev/sharedUiComponents/src/fluent/hoc/buttonLine.tsx
@@ -1,20 +1,7 @@
-import { Body1, Button as FluentButton, makeStyles, tokens } from "@fluentui/react-components";
-import { LineContainer } from "./propertyLine";
+import { LineContainer } from "./propertyLines/propertyLine";
 import type { FunctionComponent } from "react";
-import type { FluentIcon } from "@fluentui/react-icons";
-
-const useButtonLineStyles = makeStyles({
-    button: {
-        border: `1px solid ${tokens.colorBrandBackground}`,
-    },
-});
-
-export type ButtonProps = {
-    onClick: () => void;
-    icon?: FluentIcon;
-    label: string;
-    disabled?: boolean;
-};
+import { Button } from "../primitives/button";
+import type { ButtonProps } from "../primitives/button";
 
 /**
  * Wraps a button with a label in a line container
@@ -26,17 +13,5 @@ export const ButtonLine: FunctionComponent<ButtonProps> = (props) => {
         <LineContainer>
             <Button {...props} />
         </LineContainer>
-    );
-};
-
-export const Button: FunctionComponent<ButtonProps> = (props) => {
-    const classes = useButtonLineStyles();
-    // eslint-disable-next-line @typescript-eslint/naming-convention
-    const { icon: Icon, ...buttonProps } = props;
-
-    return (
-        <FluentButton iconPosition="after" className={classes.button} {...buttonProps} icon={Icon && <Icon />}>
-            <Body1>{props.label}</Body1>
-        </FluentButton>
     );
 };
