+++ packages/dev/sharedUiComponents/src/fluent/hoc/propertyLines/booleanBadgePropertyLine.tsx
@@ -8,7 +8,7 @@ import type { PropertyLineProps } from "./propertyLine";
  * @param props - The properties for the PropertyLine, including the boolean value to display.
  * @returns A PropertyLine component with a PresenceBadge indicating the boolean state.
  */
-export const ReadonlyBooleanLine: FunctionComponent<PropertyLineProps & { value: boolean }> = (props) => (
+export const BooleanBadgePropertyLine: FunctionComponent<PropertyLineProps & { value: boolean }> = (props) => (
     <PropertyLine label={props.label}>
         <PresenceBadge status={props.value ? "available" : "do-not-disturb"} outOfOffice={true} />
     </PropertyLine>
