+++ packages/dev/sharedUiComponents/src/fluent/hoc/propertyLines/checkboxPropertyLine.tsx
@@ -1,8 +1,8 @@
 import { PropertyLine } from "./propertyLine";
 import type { PropertyLineProps } from "./propertyLine";
 import type { FunctionComponent } from "react";
-import { Checkbox } from "../primitives/checkbox";
-import type { CheckboxProps } from "../primitives/checkbox";
+import { Checkbox } from "../../primitives/checkbox";
+import type { CheckboxProps } from "../../primitives/checkbox";
 
 /**
  * Wraps a checkbox in a property line
