+++ packages/dev/sharedUiComponents/src/fluent/hoc/propertyLines/colorPropertyLine.tsx
@@ -3,12 +3,12 @@ import { forwardRef, useState } from "react";
 
 import type { PropertyLineProps } from "./propertyLine";
 import { PropertyLine } from "./propertyLine";
-import { SyncedSliderLine } from "./syncedSliderLine";
+import { SyncedSliderPropertyLine } from "./syncedSliderPropertyLine";
 
 import type { Color3 } from "core/Maths/math.color";
 import { Color4 } from "core/Maths/math.color";
-import { ColorPickerPopup } from "../primitives/colorPicker";
-import type { ColorPickerProps } from "../primitives/colorPicker";
+import { ColorPickerPopup } from "../../primitives/colorPicker";
+import type { ColorPickerProps } from "../../primitives/colorPicker";
 
 export type ColorPropertyLineProps = ColorPickerProps<Color3 | Color4> & PropertyLineProps;
 
@@ -45,10 +45,10 @@ const ColorPropertyLine = forwardRef<HTMLDivElement, ColorPropertyLineProps>((pr
             {...props}
             expandedContent={
                 <>
-                    <SyncedSliderLine label="R" value={color.r * 255} min={0} max={255} onChange={(value) => onSliderChange(value, "r")} />
-                    <SyncedSliderLine label="G" value={color.g * 255} min={0} max={255} onChange={(value) => onSliderChange(value, "g")} />
-                    <SyncedSliderLine label="B" value={color.b * 255} min={0} max={255} onChange={(value) => onSliderChange(value, "b")} />
-                    {color instanceof Color4 && <SyncedSliderLine label="A" value={color.a} min={0} max={1} step={0.01} onChange={(value) => onSliderChange(value, "a")} />}
+                    <SyncedSliderPropertyLine label="R" value={color.r * 255} min={0} max={255} onChange={(value) => onSliderChange(value, "r")} />
+                    <SyncedSliderPropertyLine label="G" value={color.g * 255} min={0} max={255} onChange={(value) => onSliderChange(value, "g")} />
+                    <SyncedSliderPropertyLine label="B" value={color.b * 255} min={0} max={255} onChange={(value) => onSliderChange(value, "b")} />
+                    {color instanceof Color4 && <SyncedSliderPropertyLine label="A" value={color.a} min={0} max={1} step={0.01} onChange={(value) => onSliderChange(value, "a")} />}
                 </>
             }
         >
