+++ packages/dev/sharedUiComponents/src/fluent/hoc/propertyLines/switchPropertyLine.tsx
@@ -1,8 +1,8 @@
 import { PropertyLine } from "./propertyLine";
 import type { PropertyLineProps } from "./propertyLine";
 import type { FunctionComponent } from "react";
-import { Switch } from "../primitives/switch";
-import type { SwitchProps } from "../primitives/switch";
+import { Switch } from "../../primitives/switch";
+import type { SwitchProps } from "../../primitives/switch";
 /**
  * Wraps a switch in a property line
  * @param props - The properties for the switch and property line
