+++ packages/dev/sharedUiComponents/src/fluent/hoc/propertyLines/syncedSliderPropertyLine.tsx
@@ -1,16 +1,16 @@
 import { PropertyLine } from "./propertyLine";
 import type { PropertyLineProps } from "./propertyLine";
-import { SyncedSliderInput } from "../primitives/syncedSlider";
-import type { SyncedSliderProps } from "../primitives/syncedSlider";
+import { SyncedSliderInput } from "../../primitives/syncedSlider";
+import type { SyncedSliderProps } from "../../primitives/syncedSlider";
 import type { FunctionComponent } from "react";
 
-type SyncedSliderLineProps = SyncedSliderProps & PropertyLineProps;
+type SyncedSliderPropertyProps = SyncedSliderProps & PropertyLineProps;
 /**
  * Renders a simple wrapper around the SyncedSliderInput
  * @param props
  * @returns
  */
-export const SyncedSliderLine: FunctionComponent<SyncedSliderLineProps> = (props): React.ReactElement => {
+export const SyncedSliderPropertyLine: FunctionComponent<SyncedSliderPropertyProps> = (props): React.ReactElement => {
     const { label, description, ...sliderProps } = props;
     return (
         <PropertyLine label={label} description={description}>
