+++ packages/dev/sharedUiComponents/src/fluent/hoc/propertyLines/textPropertyLine.tsx
