+++ packages/dev/sharedUiComponents/src/split/splitContainer.tsx
@@ -1,7 +1,8 @@
 import type { PropsWithChildren } from "react";
-import { useEffect, useRef } from "react";
+import { useContext, useEffect, useRef } from "react";
 import * as styles from "./splitContainer.module.scss";
 import { ControlledSize, SplitDirection, SplitContext } from "./splitContext";
+import { ToolContext } from "shared-ui-components/fluent/hoc/fluentToolWrapper";
 
 /**
  * Split container properties
@@ -77,6 +78,7 @@ export const SplitContainer: React.FC<PropsWithChildren<ISplitContainerProps>> =
     const controllers: number[][] = [];
     const minSizes: number[] = [];
     const maxSizes: number[] = [];
+    const useFluent = useContext(ToolContext).useFluent;
 
     const buildGridDefinition = () => {
         if (!elementRef.current) {
@@ -319,6 +321,7 @@ export const SplitContainer: React.FC<PropsWithChildren<ISplitContainerProps>> =
             <div
                 id={props.id}
                 className={styles["split-container"] + " " + props.className}
+                style={useFluent ? { height: "100vh" } : undefined} // temporary until splitContainer is fluentized
                 ref={elementRef}
                 onPointerDown={(evt) => props.onPointerDown && props.onPointerDown(evt)}
                 onPointerMove={(evt) => props.onPointerMove && props.onPointerMove(evt)}
